IaC/scripts/stream-script-folder/requirements.txt
requests==2.31.0
boto3==1.26.152
python-dotenv==1.0.0